backend/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-modelcluster==6.3
django-permissionedforms==0.1
django-storages==1.14.3
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.7
l18n==2021.3
laces==0.1.1
openpyxl==3.1.2
pillow==10.3.0
pillow_heif==0.16.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
urllib3==2.2.1
wagtail==6.1
Willow==1.8.0
